Refactor Password Reset Controller
https://matijabojanic.atlassian.net/browse/OL3-120
