+++ app/Services/PasswordResetService.php
@@ -1,0 +1,32 @@
+<?php
+
+namespace App\Services;
+
+
+use App\Http\Requests\ResetPasswordRequest;
+use Illuminate\Auth\Events\PasswordReset;
+use Illuminate\Support\Facades\Hash;
+use Illuminate\Support\Facades\Password;
+use Illuminate\Support\Facades\Auth;
+
+class PasswordResetService
+{
+    public function sendResetLink(string $email): string
+    {
+        return Password::sendResetLink(['email' => $email]);
+    }
+
+    public function resetPassword(ResetPasswordRequest $request)
+    {
+        return Password::reset(
+            $request->only('email', 'password', 'password_confirmation', 'token'),
+            function ($user, $password) use ($request) {
+                $user->password = Hash::make($password);
+                $user->save();
+                event(new PasswordReset($user));
+
+                Auth::login($user);
+            }
+        );
+    }
+}
